move async API code to cheevos_client.c
